Give blobs a short term memory
Every tick, several of a blob's brain's previous outputs will be used as
inputs. Currently, there are six memory inputs and outputs. They are offset
by 1, 2, 4, 8, 16, and 32 ticks. After x number of ticks (designated by the
offsets), a output will be fed into a blob's brain as an input. Essentially,
the blob is sending off messages to itself in the future. What this data is
and how it is interpreted is completely arbitrary. Evolution will decide
how to utilize this memory. Hopefully some more intelligent behaviors will
arise as a consequence of this addition to the blob brain.